+++ lib/ui/scaffolds/bnb_home.dart
@@ -1,4 +1,4 @@
-import 'package:courses_in_english/ui/screens/location/locations_screen.dart';
+import 'package:courses_in_english/ui/screens/locations_screen.dart';
 import 'package:courses_in_english/ui/screens/course_list_screen.dart';
 import 'package:courses_in_english/ui/screens/sample_screen.dart';
 import 'package:courses_in_english/ui/screens/timetable_screen.dart';
