+++ lib/ui/scaffolds/location_detail_screen.dart
