+++ lib/ui/screens/locations_screen.dart
@@ -1,7 +1,7 @@
 import 'package:courses_in_english/connect/dataprovider/campus/campus_provider.dart';
 import 'package:courses_in_english/connect/dataprovider/data.dart';
 import 'package:courses_in_english/model/campus/campus.dart';
-import 'package:courses_in_english/ui/screens/location/location_detail_screen.dart';
+import 'package:courses_in_english/ui/scaffolds/location_detail_screen.dart';
 import 'package:flutter/material.dart';
 
 /// Location screen showing the different campuses of the munich university of applied sciences.
